setup/cosmos/requirements.txt
azure-cosmos==4.5.0
